V0.0.6b - Refactor
App.ts não tava sendo importado direito no server, commit com o git pra um nome diferente antes de voltar resolveu
