+++ src/Server.ts
@@ -1,4 +1,4 @@
-import App from "./Application";
+import App from "./App";
 import "dotenv/config";
 
 const PORT = process.env.PORT || 2020;
